v1.19.0 — perbaiki build: tipe record rekap `qty` dan `? kurang : 1`
CI menolak rilis pertama v1.19.0 (run 35957713070) di langkah
`flutter analyze lib/`, dengan dua galat tipe:

  error • A value of type 'List<({String name, int profit, double qty, int
    revenue})>' can't be returned from the method 'getRekapProduk' because it
    has a return type of 'Future<List<({String name, int profit, int qty, int
    revenue})>>' • lib/data/repositories/accounting_repository.dart:281
  error • The argument type 'num' can't be assigned to the parameter type
    'double' • lib/features/stok/stok_menipis_screen.dart:337

Keduanya persis kelas kesalahan yang tidak bisa dilihat pemeriksa mana pun di
repo ini: pemeriksa hanya membaca teks, dan tidak ada SDK Dart di mesin ini.
Tag v1.19.0 tidak pernah terbuat dan rilis pelanggan tetap v1.18.0, jadi tidak
ada yang perlu dicabut — versinya tetap 1.19.0+25.

1. Tipe record rekap `qty` dideklarasikan TIGA kali — repositori, layar
   rincian, dan pembungkus ekspor. Ketiganya masih `int` sementara badan
   metodenya sudah mengembalikan `double`. Yang dibaca pemanggil adalah
   DEKLARASINYA, jadi `return` di repositori ditolak. Ketiganya kini `double`.

2. `final saran = kurang > 0 ? kurang : 1;` bertipe `num` — gabungan `double`
   dan `int` — dan `Formatters.jumlah` menuntut `double`. Ditulis `1.0`.

Ikut diperbaiki, dan ini belum sampai CI tapi pelanggan akan melihatnya:
jumlah rekap per produk di layar rincian dan di ekspor CSV/PDF masih
diinterpolasi mentah (`'${r.qty} pcs'`), yang mencetak "2.0" untuk bilangan
bulat dan "0.5" bertitik untuk pecahan. Kini lewat `Formatters.jumlah` —
"2" dan "0,5". Pemisah CSV di aplikasi ini titik koma, jadi koma desimal aman.

Pemeriksa ikut diperkuat (check_phase_e 776 -> 786 asersi, negatif_27
81 -> 87 mutasi): ketiga deklarasi record wajib `double` dan tidak boleh ada
yang tertinggal bulat, ketiga tempat cetak wajib lewat `Formatters.jumlah`,
dan `saran` wajib memakai literal `1.0`. Tipe Dart tetap tidak bisa dilihat
pemeriksa mana pun — yang dijaga di sini adalah bentuk teksnya.

+++ lib/data/repositories/accounting_repository.dart
@@ -261,7 +261,7 @@ class AccountingRepository {
   }
 
   /// Ringkasan produk terjual per produk pada rentang waktu.
-  Future<List<({String name, int qty, int revenue, int profit})>>
+  Future<List<({String name, double qty, int revenue, int profit})>>
       getRekapProduk(DateTime start, DateTime end) async {
     final db = await _db.database;
     final rows = await db.rawQuery('''
